Add screenshot and thumbnail functionality
Introduces --screenshot, --fullPageScreenshot, and --thumbnail as
separate flags which can be combined as desired.

Thumbnails are the same size in terms of height and width as the other
screenshots due to limitations in the Puppeteer API but are taken as
JPEGs at 75% quality, which are significantly smaller than the pngs.

At this stage, screenshots are not yet written into WACZ files.